+++ src/importlinter/domain/imports.py
@@ -83,24 +83,41 @@ class DirectImport(ValueObject):
         return hash((str(self), self.line_contents))
 
 
+class ModuleExpression(ValueObject):
+    """
+    A user-submitted expression describing a module or a set of modules.
+
+    Sets of modules are notated using * or ** wildcards.
+    Examples:
+        "mypackage.foo.bar": a single module
+        "mypackage.foo.*": all direct submodules in the foo subpackage
+        "mypackage.*.bar": all bar-submodules of any mypackage submodule
+        "mypackage.**": all modules in the mypackage package
+
+    Note that * and ** cannot be mixed in the same expression.
+    """
+
+    def __init__(self, expression: str) -> None:
+        self.expression = expression
+
+    def has_wildcard_expression(self) -> bool:
+        return "*" in self.expression
+
+
 class ImportExpression(ValueObject):
     """
     A user-submitted expression describing an import or set of imports.
 
-    Sets of imports are notated using * wildcards.
-    These wildcards can stand in for a module name or part of a name, but they do
-    not extend to subpackages.
-
-    For example, "mypackage.*" refers to every child subpackage of mypackage.
-    It does not, however, include more distant descendants such as mypackage.foo.bar.
+    The importer and imported expressions are both ModuleExpressions
+    (see ModuleExpression for details).
     """
 
-    def __init__(self, importer: str, imported: str) -> None:
+    def __init__(self, importer: ModuleExpression, imported: ModuleExpression) -> None:
         self.importer = importer
         self.imported = imported
 
     def has_wildcard_expression(self) -> bool:
-        return "*" in self.imported or "*" in self.importer
+        return self.imported.has_wildcard_expression() or self.importer.has_wildcard_expression()
 
     def __str__(self) -> str:
-        return "{} -> {}".format(self.importer, self.imported)
+        return "{} -> {}".format(self.importer.expression, self.imported.expression)