+++ tests/unit/contracts/test_forbidden.py
@@ -60,6 +60,19 @@ class TestForbiddenContract:
                     ],
                 },
                 {
+                    "upstream_module": "mypackage.green",
+                    "downstream_module": "mypackage.three",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.three",
+                                "imported": "mypackage.green",
+                                "line_numbers": (4,),
+                            }
+                        ]
+                    ],
+                },
+                {
                     "upstream_module": "mypackage.purple",
                     "downstream_module": "mypackage.two",
                     "chains": [
@@ -74,19 +87,6 @@ class TestForbiddenContract:
                                 "imported": "mypackage.purple",
                                 "line_numbers": (1,),
                             },
-                        ]
-                    ],
-                },
-                {
-                    "upstream_module": "mypackage.green",
-                    "downstream_module": "mypackage.three",
-                    "chains": [
-                        [
-                            {
-                                "importer": "mypackage.three",
-                                "imported": "mypackage.green",
-                                "line_numbers": (4,),
-                            }
                         ]
                     ],
                 },
@@ -152,6 +152,168 @@ class TestForbiddenContract:
 
         check = contract.check(graph=graph, verbose=False)
         assert check.kept
+
+    def test_wildcards_in_source_modules_are_resolved(self):
+        graph = self._build_graph()
+        contract = self._build_contract(
+            forbidden_modules=("mypackage.green"),
+            source_modules=("mypackage.one.*",),
+            include_external_packages=False,
+        )
+
+        check = contract.check(graph=graph, verbose=False)
+        assert check.metadata == {
+            "invalid_chains": [
+                {
+                    "upstream_module": "mypackage.green",
+                    "downstream_module": "mypackage.one.alpha",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha",
+                                "imported": "mypackage.green.beta",
+                                "line_numbers": (3,),
+                            },
+                        ],
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+            ],
+        }
+
+    def test_recursive_wildcards_in_source_modules_are_resolved(self):
+        graph = self._build_graph()
+        contract = self._build_contract(
+            forbidden_modules=("mypackage.green"),
+            source_modules=("mypackage.one.**",),
+            include_external_packages=False,
+        )
+
+        check = contract.check(graph=graph, verbose=False)
+        assert check.metadata == {
+            "invalid_chains": [
+                {
+                    "upstream_module": "mypackage.green",
+                    "downstream_module": "mypackage.one.alpha",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha",
+                                "imported": "mypackage.green.beta",
+                                "line_numbers": (3,),
+                            },
+                        ],
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+                {
+                    "upstream_module": "mypackage.green",
+                    "downstream_module": "mypackage.one.alpha.circle",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+            ],
+        }
+
+    def test_wildcards_in_forbidden_modules_are_resolved(self):
+        graph = self._build_graph()
+        contract = self._build_contract(
+            forbidden_modules=("mypackage.green.*"),
+            source_modules=("mypackage.one",),
+            include_external_packages=False,
+        )
+
+        check = contract.check(graph=graph, verbose=False)
+        assert check.metadata == {
+            "invalid_chains": [
+                {
+                    "upstream_module": "mypackage.green.beta",
+                    "downstream_module": "mypackage.one",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha",
+                                "imported": "mypackage.green.beta",
+                                "line_numbers": (3,),
+                            },
+                        ],
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+            ],
+        }
+
+    def test_recursive_wildcards_in_forbidden_modules_are_resolved(self):
+        graph = self._build_graph()
+        contract = self._build_contract(
+            forbidden_modules=("mypackage.green.**"),
+            source_modules=("mypackage.one",),
+            include_external_packages=False,
+        )
+
+        check = contract.check(graph=graph, verbose=False)
+        assert check.metadata == {
+            "invalid_chains": [
+                {
+                    "upstream_module": "mypackage.green.beta",
+                    "downstream_module": "mypackage.one",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha",
+                                "imported": "mypackage.green.beta",
+                                "line_numbers": (3,),
+                            },
+                        ],
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+                {
+                    "upstream_module": "mypackage.green.beta.sphere",
+                    "downstream_module": "mypackage.one",
+                    "chains": [
+                        [
+                            {
+                                "importer": "mypackage.one.alpha.circle",
+                                "imported": "mypackage.green.beta.sphere",
+                                "line_numbers": (8,),
+                            },
+                        ],
+                    ],
+                },
+            ],
+        }
 
     def test_ignore_imports_with_wildcards(self):
         graph = self._build_graph()
@@ -415,13 +577,15 @@ class TestForbiddenContract:
         ignore_imports=None,
         include_external_packages=False,
         allow_indirect_imports=None,
+        source_modules=None,
     ):
         session_options = {"root_packages": ["mypackage"]}
         if include_external_packages:
             session_options["include_external_packages"] = "True"
 
         contract_options = {
-            "source_modules": ("mypackage.one", "mypackage.two", "mypackage.three"),
+            "source_modules": source_modules
+            or ("mypackage.one", "mypackage.two", "mypackage.three"),
             "forbidden_modules": forbidden_modules,
             "ignore_imports": ignore_imports or [],
         }
@@ -639,25 +803,25 @@ class TestVerbosePrint:
             Found 0 illegal chains in 10s.
             Searching for import chains from mypackage.one to mypackage.green...
             Found 1 illegal chain in 10s.
+            Searching for import chains from mypackage.one to mypackage.purple...
+            Found 0 illegal chains in 10s.
             Searching for import chains from mypackage.one to mypackage.yellow...
             Found 0 illegal chains in 10s.
-            Searching for import chains from mypackage.one to mypackage.purple...
+            Searching for import chains from mypackage.three to mypackage.blue...
+            Found 0 illegal chains in 10s.
+            Searching for import chains from mypackage.three to mypackage.green...
+            Found 1 illegal chain in 10s.
+            Searching for import chains from mypackage.three to mypackage.purple...
+            Found 0 illegal chains in 10s.
+            Searching for import chains from mypackage.three to mypackage.yellow...
             Found 0 illegal chains in 10s.
             Searching for import chains from mypackage.two to mypackage.blue...
             Found 0 illegal chains in 10s.
             Searching for import chains from mypackage.two to mypackage.green...
             Found 0 illegal chains in 10s.
-            Searching for import chains from mypackage.two to mypackage.yellow...
-            Found 0 illegal chains in 10s.
             Searching for import chains from mypackage.two to mypackage.purple...
             Found 1 illegal chain in 10s.
-            Searching for import chains from mypackage.three to mypackage.blue...
-            Found 0 illegal chains in 10s.
-            Searching for import chains from mypackage.three to mypackage.green...
-            Found 1 illegal chain in 10s.
-            Searching for import chains from mypackage.three to mypackage.yellow...
-            Found 0 illegal chains in 10s.
-            Searching for import chains from mypackage.three to mypackage.purple...
+            Searching for import chains from mypackage.two to mypackage.yellow...
             Found 0 illegal chains in 10s.
             """
         )